Rename custom retrofit adapter
