--- omisego-sdk/src/main/java/co/omisego/omisego/custom/retrofit2/adapter/OMGAdapter.kt
@@ -1,25 +1,0 @@
-package co.omisego.omisego.custom.retrofit2.adapter
-
-/**
- * OmiseGO
- *
- * Created by Phuchit Sirimongkolsathien on 9/3/2018 AD.
- * Copyright © 2017-2018 OmiseGO. All rights reserved.
- */
-
-import co.omisego.omisego.model.OMGResponse
-import co.omisego.omisego.custom.retrofit2.adapter.OMGCallAdapter
-import com.google.gson.TypeAdapter
-import retrofit2.Call
-import retrofit2.CallAdapter
-import java.lang.reflect.Type
-
-internal class OMGAdapter<R>(private val responseType: Type, private val adapter: TypeAdapter<R>) : CallAdapter<OMGResponse<R>, OMGCall<OMGResponse<R>>> {
-    override fun adapt(call: Call<OMGResponse<R>>): OMGCall<OMGResponse<R>> {
-        return OMGCallAdapter(call)
-    }
-
-    override fun responseType(): Type {
-        return responseType
-    }
-}
